♻️  (#444) Router 구조 리팩토링 (경로 상수 및 헬퍼 분리)
